Separate code for selection of remote Uri from updating settings (#433)
